chore: move verificationAbi to constants
