+++ src/lib/constants/verificationAbi.ts
